+++ lib/ransack/configuration.rb
@@ -34,7 +34,8 @@ module Ransack
       :down_arrow => '&#9650;'.freeze,
       :default_arrow => nil,
       :remove_association_no_negative_assoc => true,
-      :sanitize_scope_args => true
+      :sanitize_scope_args => true,
+      :postgres_fields_sort_option => nil
     }
 
     def configure
@@ -140,6 +141,21 @@ module Ransack
     #
     def sanitize_custom_scope_booleans=(boolean)
       self.options[:sanitize_scope_args] = boolean
+    end
+
+    # The `NULLS FIRST` and `NULLS LAST` options can be used to determine
+    # whether nulls appear before or after non-null values in the sort ordering.
+    #
+    # User may want to configure it like this:
+    #
+    # Ransack.configure do |c|
+    #   c.postgres_fields_sort_option = :nulls_first # or :nulls_last
+    # end
+    #
+    # See this feature: https://www.postgresql.org/docs/13/queries-order.html
+    #
+    def postgres_fields_sort_option=(setting)
+      self.options[:postgres_fields_sort_option] = setting
     end
 
     # By default, Ransack displays sort order indicator arrows in sort links.